requirements.txt
python-dotenv
python_fasthtml
pytz
supabase